+++ config/mode.ts
@@ -31,9 +31,10 @@ export default <NetworkData>{
         excludedTokenAddresses: [],
     },
     rpcUrl: env.DRPC_API_KEY
-        ? `https://direct.drpc.org/ogrpc?network=mode&dkey=${env.DRPC_API_KEY}`
+        ? `https://lb.drpc.org/ogrpc?network=mode&dkey=${env.DRPC_API_KEY}`
         : 'https://mainnet.mode.network',
-    rpcMaxBlockRange: 10000,
+    rpcMaxBlockRange: 5000,
+    acceptableSGLag: 30, // ~1min
     protocolToken: 'bal',
     bal: {
         address: '0xd08a2917653d4e460893203471f0000826fb4034',