+++ config/sepolia.ts
@@ -34,9 +34,10 @@ export default <NetworkData>{
         excludedTokenAddresses: [],
     },
     rpcUrl: env.DRPC_API_KEY
-        ? `https://direct.drpc.org/ogrpc?network=sepolia&dkey=${env.DRPC_API_KEY}`
+        ? `https://lb.drpc.org/ogrpc?network=sepolia&dkey=${env.DRPC_API_KEY}`
         : 'https://gateway.tenderly.co/public/sepolia',
     rpcMaxBlockRange: 10000,
+    acceptableSGLag: 100, // ~1min
     protocolToken: 'bal',
     bal: {
         address: '0xb19382073c7A0aDdbb56Ac6AF1808Fa49e377B75',