+++ modules/actions/cow-amm/index.ts
@@ -1,6 +1,4 @@
-export * from './sync-pools';
 export * from './upsert-pools';
-export * from './fetch-new-pools';
 export * from './fetch-changed-pools';
 export * from './sync-swaps';
 export * from './sync-join-exits';