+++ modules/sources/logs/index.ts
@@ -1,3 +1,2 @@
 export * from './get-changed-pools';
 export * from './get-changed-pools-v2';
-export * from './get-changed-cow-amm-pools';