+++ modules/sources/subgraphs/joined-client.ts
@@ -1,7 +1,6 @@
-import { getVaultSubgraphClient } from './balancer-v3-vault';
-import { getPoolsSubgraphClient } from './balancer-v3-pools';
+import { V3VaultSubgraphClient } from './balancer-v3-vault';
+import { V3PoolsSubgraphClient } from './balancer-v3-pools';
 import { PoolsQueryVariables } from './balancer-v3-vault/generated/types';
-import { Chain } from '@prisma/client';
 
 export type V3JoinedSubgraphClient = ReturnType<typeof getV3JoinedSubgraphClient>;
 
@@ -11,11 +10,23 @@ export type V3JoinedSubgraphPool = ReturnType<V3JoinedSubgraphClient['getAllInit
     ? T
     : never;
 
-export const getV3JoinedSubgraphClient = (vaultSubgraphUrl: string, poolsSubgraphUrl: string, chain: Chain) => {
-    const vaultSubgraphClient = getVaultSubgraphClient(vaultSubgraphUrl, chain);
-    const poolsSubgraphClient = getPoolsSubgraphClient(poolsSubgraphUrl, chain);
-
+export const getV3JoinedSubgraphClient = (
+    vaultSubgraphClient: V3VaultSubgraphClient,
+    poolsSubgraphClient: V3PoolsSubgraphClient,
+) => {
     return {
+        getMetadata: async () => {
+            const vault = await vaultSubgraphClient.getMetadata();
+            const pools = await poolsSubgraphClient.getMetadata();
+
+            return {
+                ...vault,
+                block: {
+                    ...vault.block,
+                    number: Math.min(vault.block.number, pools.block.number),
+                },
+            };
+        },
         getAllInitializedPools: async (where?: PoolsQueryVariables['where']) => {
             const vaultPools = await vaultSubgraphClient.getAllInitializedPools(where);
             const vaultPoolsMap = vaultPools.reduce((acc, pool) => {